host: Move several items from host to emulator:
* init
* handle_events
* screen_render.cpp and screen_render.h
* app.cpp and app.h
* config functions
* discrete GPU exports
* emulator-called SFO functions
