+++ src/emulator/host/include/host/config.h
@@ -40,36 +40,3 @@ struct Config {
     optional<std::string> pref_path;
     bool archive_log = false;
 };
-
-namespace config {
-
-enum class InitResult {
-    OK,
-    QUIT,
-    INCORRECT_ARGS,
-};
-
-/*
- * \brief Initializes config system from a YML file.
- * 
- * \param cfg Config options are returned via this parameter.
- * \return True on loading success.
-*/
-bool deserialize(Config &cfg);
-
-/*
- * \brief Save emulator config to a YML file.
- *
- * \return True on saving success.
-*/
-bool serialize(Config &cfg);
-
-/**
-  * \brief Initializes config system, parsing command-line args and handling some basic ones:
-  *        --help, --version, --log-level
-  * \param cfg Config options are returned via this parameter.
-  * \return True on success, false on error.
-  */
-InitResult init(Config &cfg, int argc, char **argv);
-
-} // namespace config
