+++ src/emulator/main.cpp
@@ -15,14 +15,13 @@
 // with this program; if not, write to the Free Software Foundation, Inc.,
 // 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
+#include "app.h"
+#include "config.h"
+#include "screen_render.h"
+
 #include <gui/functions.h>
-#include <host/app.h>
-#include <host/config.h>
-#include <host/functions.h>
-#include <host/screen_render.h>
 #include <host/state.h>
 #include <host/version.h>
-#include <kernel/thread/thread_functions.h>
 #include <shader/spirv_recompiler.h>
 #include <util/log.h>
 #include <util/string_utils.h>
@@ -30,8 +29,6 @@
 #include <SDL.h>
 
 #include <cstdlib>
-#include <iterator>
-#include <utility>
 
 int main(int argc, char *argv[]) {
     logging::init();
