+++ src/emulator/screen_render.cpp
@@ -15,7 +15,8 @@
 // with this program; if not, write to the Free Software Foundation, Inc.,
 // 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-#include <host/screen_render.h>
+#include "screen_render.h"
+
 #include <host/state.h>
 #include <util/log.h>
