+++ src/emulator/screen_render.h
